backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
Pillow==10.2.0
python-magic==0.4.27
django-storages==1.14.2
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0